feat: add `throwWhenDisposed` option to cart
This commit introduces a new `throwWhenDisposed` boolean option to `BehaviorOptions`.

When set to `true` (default), the cart will throw a `CartDisposedException` if an attempt is made to modify it after it has been disposed.
When set to `false`, such attempts will be silently ignored.

The `checkIfDisposed` method in `CartChangeNotifierDisposeMixin` has been updated to respect this new option.

A new test case has been added in `cart_disposed_test.dart` to verify the behavior when `throwWhenDisposed` is `false`.